Massive refactoring to improve maintainability.
