+++ utils/dbToRes.js
@@ -1,3 +1,4 @@
+const convertCase = require('./convertCase');
 //Convets snake_case to camelCase
 const toCamel = string => {
   const result = string.split("_").map((part, i) => {
@@ -11,13 +12,7 @@ const toCamel = string => {
 
 //Converts key in key-object pairs to camelCase
 const dbToRes = obj => {
-  const cameled = {};
-  Object.entries(obj).forEach(item => {
-    const [key, value] = item;
-    cameled[toCamel(key)] = value;
-  });
-
-  return cameled;
+  return convertCase(toCamel, obj);
 };
 
 module.exports = dbToRes;
