+++ utils/reqToDb.js
@@ -1,3 +1,4 @@
+const convertCase = require('./convertCase');
 //Converts camelCase to snake_case
 const toSnake = string => {
   const capitalPattern = /[a-z]/;
@@ -12,13 +13,7 @@ const toSnake = string => {
 
 //Converts key in key-object pairs to snake_case
 const reqToDb = obj => {
-  const snaked = {};
-  Object.entries(obj).forEach(item => {
-    const [key, value] = item;
-    snaked[toSnake(key)] = value;
-  });
-
-  return snaked;
+  return convertCase(toSnake, obj);
 };
 
 module.exports = reqToDb;
